+++ scripts/effects/kerning.lua
@@ -25,7 +25,8 @@ local xform_rotation_x = 0.0 --track@xform_rotation_x:Transform::Rotation::X,-36
 local xform_rotation_y = 0.0 --track@xform_rotation_y:Transform::Rotation::Y,-3600,3600,0,0.01
 local xform_rotation_z = 0.0 --track@xform_rotation_z:Transform::Rotation::Z,-3600,3600,0,0.01
 --#define EULER XYZ Euler=5,XZY Euler=7,YXZ Euler=11,YZX Euler=15,ZXY Euler=19,ZYX Euler=21
-local xform_rotation_mode = 21 --select@xform_rotation_mode:Transform::Rotation::Mode=21,Quaternion=0,Axis Angle=1,${EULER}
+--#define ROTATION_MODES Quaternion=0,Axis Angle=1,${EULER}
+local xform_rotation_mode = 21 --select@xform_rotation_mode:Transform::Rotation::Mode=21,${ROTATION_MODES}
 --trackgroup@xform_rotation_x,xform_rotation_y,xform_rotation_z:Group::Transform::Rotation
 --separator:Scale
 local xform_scale_x = 100.0 --track@xform_scale_x:Transform::Scale::X,-10000,10000,100,0.01
@@ -38,7 +39,8 @@ local xform_scale_z = 100.0 --track@xform_scale_z:Transform::Scale::Z,-10000,100
 --#define CONTRAST Overlay=5,Linear Light=11
 --#define COMPARATIVE Difference=12,Subtract=2
 --#define HSL Color=9,Luminosity=8
-local xform_blend_mode = 0 --select@xform_blend_mode:Transform::Compositing::Blend Mode,Normal=0,${DARKEN},${LIGHTEN},${CONTRAST},${COMPARATIVE},${HSL}
+--#define BLEND_MODES Normal=0,${DARKEN},${LIGHTEN},${CONTRAST},${COMPARATIVE},${HSL}
+local xform_blend_mode = 0 --select@xform_blend_mode:Transform::Compositing::Blend Mode,${BLEND_MODES}
 local xform_opacity = 100.0 --track@xform_opacity:Transform::Compositing::Opacity,0,100,100,0.01
 --separator:Target
 local xform_target_local_space = true --checksection@xform_target_local_space:Transform::Target::Local Space,true
@@ -78,7 +80,9 @@ do
     xform_scale_y = xform_scale_y * 0.01
     xform_scale_z = xform_scale_z * 0.01
     xform_opacity = xform_opacity * 0.01
+
     tint_opacity = tint_opacity * 0.01
+
     influence = influence * 0.01
 
     local text = getvalue("テキスト", "テキスト")